requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
python-dotenv==1.0.0
Werkzeug==2.3.7
reportlab==4.0.4
openpyxl==3.1.2
gunicorn==24.1.1